Make X-version of emacs past to current cursor location instead of first moving cursor
I.e. before when clicking the third mouse button in X, it would move
the emacs cursor to the mouse cursor location on screen, and then
paste. Now instead it behaves as if running emacs -nw in a terminal
emulator, and pastes the text to the emacs cursor location without
moving it anywhere.